pggan-pytorch/requirements.txt
librosa==0.4.3
numpy==1.13.0
scipy==0.19.0
SoundFile==0.9.0.post1
Pillow==4.0.0
torch==0.2.0.post3
